+++ NEWS.md
@@ -1,6 +1,6 @@
 # googleCloudRunner 0.1.1.9000
 
-* ...
+* Add `port` argument to Cloud Run deployments via `cr_buildstep_run()`
 
 # googleCloudRunner 0.1.1